\name{MonophylyData}
\docType{data}
\alias{MonophylyData}
\alias{Ericactree}
\alias{Ericactribes}
\title{
Example dataset for the package MonoPhy.
}
\description{
Example dataset for the package MonoPhy. One phylogeny and one custom taxonomy file.
}
\usage{
data(Ericactree)
data(Ericactribes)
}
\details{
Ericactree is a phylogeny of the angiosperm family Ericaceae, as presented in Schwery et al. (2014), pruned to 77 species. Ericactribes is an example file for how to specify which taxonomic groups should be tested for monophyly (instead of genus names taken from the tip labels of the tree), here exemplified by tribes.
}
\references{
Schwery, O., Onstein, R.E., Bouchenak-Khelladi, Y., Xing, Y., Carter, R.J. and Linder, H.P. (2014), As old as the mountains: the radiations of the Ericaceae. New Phytologist. doi: 10.1111/nph.13234
}
\keyword{datasets}
